network_segment.rb: many changes/fixes
 'cuz Chris didn't understand CIDR's and the IPAddr class well enough back then.

cidr is no longer an attribute of a NetworkSegment, since NetworkSegments can be any arbitrary IP range, not just those that define IP subnets. As such the ::subnets class method now is just an alias for ::ip_ranges, which returns all network segments in the JSS as Range objects of IPAddr objects..

You can still create and modify NetworkSegment objects with :cidr or :mask, but they are only used to determine the ending address based on the starting address. If you know an exact ending address, use it.

equality of NetworkSegments is determined by comparing their IPAddr Ranges.