User ldap_server requires id, not just name
Changes to the Jamf Classic API have made it so that the ldap_server associated with a User cannot be identified by name alone, but instead must be identified using id. ldap_server id is now an integral part of the User object. Issue was identified on 3/24/2020 as a breaking circumstance for jssmaint —sync-ldap-people.