Fix broken config profile payload updates
Prior to this change, the behaviour was to upload the XML version of
the new profile content, but this wouldn't include the necessary
"header" information for a custom config profile and would result in
a 409 (conflict) HTTP error.

Now the uploaded payload includes the required pre-amble and the payload
contents can be updated as intended.