Handle missing properties object during OAPI generation
In Jamf Pro 11.5.1 (at least, it's probably been this way for a while)
some objects don't have a `properties` field, like the
LogOutUserCommand.

To see this, go to yourserver.jamfcloud.com/api/schema and search for
the `LogOutUserCommand` definition.

This is what it looks like in 11.5.1 and 11.6.0

```
      "LogOutUserCommand" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/MdmCommandBase"
        }, {
          "type" : "object"
        } ]
      },
```

Weirdly, the end of the array had a comment suggesting that this change
was once the behaviour. I couldn't find anything in the git history as
to why it was changed.